fix: MyStudyCard 수정
탈퇴 버튼을 클릭하면 스터디 룸으로 이동하는 버그가 있어 컴포넌트를 분리하고 스타일을 수정
e.stopPropagation을 해도 기본 a, button의 동작은 막지 않음
