+++ frontend/src/pages/my-study-page/components/my-study-card-list-section/MyStudyCardListSection.style.tsx
@@ -1,3 +1,4 @@
+import { css } from '@emotion/react';
 import styled from '@emotion/styled';
 
 import { mqDown } from '@utils';
@@ -7,6 +8,10 @@ export const SectionTitle = styled.h3`
 
   font-size: 24px;
   font-weight: 700;
+`;
+
+export const MyStudyCardListSection = styled.section`
+  padding: 8px;
 `;
 
 export const MyStudyList = styled.ul`
@@ -27,6 +32,27 @@ export const MyStudyList = styled.ul`
   }
 `;
 
-export const MyStudyCardListSection = styled.section`
-  padding: 8px;
+export const MyStudyCardItem = styled.li`
+  position: relative;
+`;
+
+export const TrashButton = styled.button`
+  ${({ theme }) => css`
+    position: absolute;
+    bottom: 12px;
+    right: 12px;
+
+    background-color: transparent;
+    border: none;
+    outline: none;
+
+    & > svg {
+      stroke: ${theme.colors.primary.base};
+
+      &:hover,
+      &:active {
+        stroke: ${theme.colors.primary.light};
+      }
+    }
+  `};
 `;
