+++ frontend/src/pages/my-study-page/components/my-study-card/MyStudyCard.style.tsx
@@ -81,20 +81,3 @@ export const MyStudyCard = styled.div<Pick<MyStudyCardProps, 'disabled'>>`
     ${disabled && disabledStyle(theme)}
   `}
 `;
-
-export const TrashButton = styled.button`
-  ${({ theme }) => css`
-    background-color: transparent;
-    border: none;
-    outline: none;
-
-    & > svg {
-      stroke: ${theme.colors.primary.base};
-
-      &:hover,
-      &:active {
-        stroke: ${theme.colors.primary.dark};
-      }
-    }
-  `};
-`;
