+++ frontend/src/pages/my-study-page/components/my-study-card/MyStudyCard.tsx
@@ -31,25 +31,6 @@ const TbCrown = () => (
   </svg>
 );
 
-const HiOutlineTrash = () => (
-  <svg
-    stroke="currentColor"
-    fill="none"
-    strokeWidth="0"
-    viewBox="0 0 24 24"
-    height="20"
-    width="20"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      strokeWidth="2"
-      d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16"
-    ></path>
-  </svg>
-);
-
 const MyStudyCard: React.FC<OptionalMyStudyCardProps> = ({
   title,
   ownerName,
@@ -77,9 +58,6 @@ const MyStudyCard: React.FC<OptionalMyStudyCardProps> = ({
           <S.Period>
             <span>{startDate}</span> ~ <span>{endDate || ''}</span>
           </S.Period>
-          <S.TrashButton>
-            <HiOutlineTrash />
-          </S.TrashButton>
         </S.Bottom>
       </S.Container>
     </S.MyStudyCard>
